# Vectorize `remove_copy` and `unique_copy`