--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -298,6 +298,17 @@ constexpr bool _Vector_alg_in_ranges_replace_is_safe =
 // Can we activate the vector algorithms for unique?
 template <class _Iter, class _Pr>
 constexpr bool _Vector_alg_in_unique_is_safe = _Equal_memcmp_is_safe<_Iter, _Iter, _Pr>;
+
+// Can use this output iterator for 'remove_copy' or 'unique_copy'?
+template <class _Out, class _In>
+constexpr bool _Ouput_iterator_for_vector_alg_is_safe() {
+    if constexpr (_Iterator_is_contiguous<_Out>) {
+        return is_same_v<_Iter_value_t<_Out>, remove_const_t<_Iter_value_t<_In>>>;
+    } else {
+        return false;
+    }
+}
+
 _STD_END
 #endif // _USE_STD_VECTOR_ALGORITHMS
 
@@ -4681,8 +4692,8 @@ _CONSTEXPR20 _OutIt remove_copy(_InIt _First, _InIt _Last, _OutIt _Dest, const _
     auto _UDest       = _STD _Get_unwrapped_unverified(_Dest);
 
 #if _USE_STD_VECTOR_ALGORITHMS
-    if constexpr (_Vector_alg_in_find_is_safe<decltype(_UFirst), _Ty> && _Iterator_is_contiguous<decltype(_UDest)>
-                  && is_same_v<remove_const_t<_Iter_value_t<decltype(_UFirst)>>, _Iter_value_t<decltype(_UDest)>>) {
+    if constexpr (_Vector_alg_in_find_is_safe<decltype(_UFirst), _Ty>
+                  && _Ouput_iterator_for_vector_alg_is_safe<decltype(_UDest), decltype(_UFirst)>()) {
         if (!_STD _Is_constant_evaluated()) {
             if (!_STD _Could_compare_equal_to_value_type<decltype(_UFirst)>(_Val)) {
                 _UDest += _ULast - _UFirst;
@@ -4932,8 +4943,7 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(indirect_binary_predicate<ranges::equal_to, projected<_It, _Pj>, const _Ty*>);
 
 #if _USE_STD_VECTOR_ALGORITHMS
-            if constexpr (_Vector_alg_in_find_is_safe<_It, _Ty> && _Iterator_is_contiguous<_Out>
-                          && is_same_v<remove_const_t<_Iter_value_t<_It>>, _Iter_value_t<_Out>>
+            if constexpr (_Vector_alg_in_find_is_safe<_It, _Ty> && _Ouput_iterator_for_vector_alg_is_safe<_Out, _It>()
                           && sized_sentinel_for<_Se, _It> && is_same_v<_Pj, identity>) {
                 if (!_STD is_constant_evaluated()) {
                     const auto _Size = _Last - _First;
@@ -5204,8 +5214,8 @@ _CONSTEXPR20 _OutIt unique_copy(_InIt _First, _InIt _Last, _OutIt _Dest, _Pr _Pr
     auto _UDest = _STD _Get_unwrapped_unverified(_Dest);
 
 #if _USE_STD_VECTOR_ALGORITHMS
-    if constexpr (_Vector_alg_in_unique_is_safe<decltype(_UFirst), _Pr> && _Iterator_is_contiguous<decltype(_UDest)>
-                  && is_same_v<remove_const_t<_Iter_value_t<decltype(_UFirst)>>, _Iter_value_t<decltype(_UDest)>>) {
+    if constexpr (_Vector_alg_in_unique_is_safe<decltype(_UFirst), _Pr>
+                  && _Ouput_iterator_for_vector_alg_is_safe<decltype(_UDest), decltype(_UFirst)>()) {
         if (!_STD _Is_constant_evaluated()) {
             const auto _First_ptr = _STD _To_address(_UFirst);
             const auto _Dest_ptr  = _STD _To_address(_UDest);
@@ -5352,8 +5362,8 @@ namespace ranges {
 
 #if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (is_same_v<_Pj, identity> && sized_sentinel_for<_Se, _It>
-                          && _Vector_alg_in_unique_is_safe<_It, _Pr> && _Iterator_is_contiguous<_Out>
-                          && is_same_v<remove_const_t<_Iter_value_t<_It>>, _Iter_value_t<_Out>>) {
+                          && _Vector_alg_in_unique_is_safe<_It, _Pr>
+                          && _Ouput_iterator_for_vector_alg_is_safe<_Out, _It>()) {
                 if (!_STD is_constant_evaluated()) {
                     const auto _Size       = _Last - _First;
                     const auto _First_ptr  = _STD to_address(_First);