@AlexGuteniev 2025-03-23
> The usual codegen gremlins that cause results variation are observed for non-vectorized tight loops. I've marked the most notorious one with clown. I can't explain that anomality.

Very wild guess: the branch forward to skip the repeated element has address exactly 0x10 lower than the main loop backward jump, so the former somehow interferes branch prediction of the latter.

![image](https://github.com/user-attachments/assets/cf6d1e77-a3fd-4029-b730-f5958e5dd908)