@StephanTLavavej 2025-04-22
Thanks! :heart_eyes_cat: I pushed fixes for significant bugs in the "can compare equal to value type" codepaths, and expanded the test coverage (auditing all existing codepaths, and verifying that the new tests caught the bugs). Please double-check.