@StephanTLavavej 2025-04-22
I resolved a trivial adjacent-add conflict with #5352 in `<algorithm>`.